third_party/blink/web_tests/http/tests/xmlhttprequest/reopen-encoding-expected.txt

This test verifies that XMLHttpRequest properly handles Content-Type when it is reused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS req.getResponseHeader('Content-Type') is "text/plain; charset=utf-8"
PASS req.responseText is "Проверка"
PASS req.getResponseHeader('Content-Type') is "text/plain; charset=windows-1251"
PASS req.responseText is "Проверка"
PASS successfullyParsed is true

TEST COMPLETE
